provided custom scroll view to weather screen + refactor
