+++ lib/features/weather/presentation/constants.dart
@@ -10,3 +10,9 @@ const weatherDescriptionTextStyle = TextStyle(
 const weatherCardBodyTextStyle = TextStyle(
   fontSize: 50,
 );
+
+const String weatherCardBackgroundImage =
+    'https://voxukraine.org/wp-content/uploads/2023/03/Depositphotos_188007782_L-1.jpg';
+
+const String weatherScreenBackgroundImage =
+    'https://i.pinimg.com/originals/4e/b5/be/4eb5be4deef02f33df2d9e807994e31c.jpg';
